% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_complete.R
\name{bayes_complete}
\alias{bayes_complete}
\title{Posterior parameters for infection and removal rates given complete data}
\usage{
bayes_complete(
  removals,
  infections,
  population_size,
  beta_rate = 1e-05,
  beta_shape = 0.001,
  gamma_rate = 0.001,
  gamma_shape = 0.001,
  num_iter = 10000,
  lag = 0
)
}
\arguments{
\item{removals}{numeric vector: removal times}

\item{infections}{numeric vector: infection times}

\item{population_size}{integer: population size}

\item{beta_rate}{numeric}

\item{beta_shape}{numeric}

\item{gamma_rate}{numeric}

\item{gamma_shape}{numeric}

\item{num_iter}{numeric}

\item{lag}{numeric: fixed exposure period}
}
\value{
numeric list of exact posterior and prior parameters
}
\description{
Parameters for independent Gibbs sampling of the infection and removal rates.
You can sample from \code{rgamma()} with the posterior parameters to get the posterior distribution.
For the infection rate, you should scale the result of \code{rgamma()} by the population size.
\code{infection.rate.samples} and \code{removal.rate.samples} contain the posterior samples.
}
