% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_complete_data.R
\name{bayes_complete_data}
\alias{bayes_complete_data}
\title{Posterior samples for infection and removal rates given complete data}
\usage{
bayes_complete_data(r, i, N)
}
\arguments{
\item{r}{numeric vector: removal times}

\item{i}{numeric vector: infection times}

\item{N}{integer: population size}
}
\value{
numeric list of posterior (infection.rate, removal.rate)
}
\description{
Gibbs sample the infection and removal rates
}
