% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_complete_multitype.R
\name{bayes_complete_multitype}
\alias{bayes_complete_multitype}
\title{Posterior sampling for multi-type infection and removal rates given complete data}
\usage{
bayes_complete_multitype(
  removals,
  infections,
  removal_classes,
  infection_classes,
  infection_class_sizes,
  beta_init,
  beta_shape,
  gamma_init,
  gamma_shape,
  num_iter = 10000,
  num_renewals = 1,
  lag = 0
)
}
\arguments{
\item{removals}{numeric vector: removal times}

\item{infections}{numeric vector: infection times}

\item{beta_init}{numeric: initial infection rate estimates}

\item{beta_shape}{numeric: shape of gamma distribution prior}

\item{gamma_init}{numeric: initial removal rate estimates}

\item{gamma_shape}{numeric: shape of gamma distribution prior}

\item{num_iter}{numeric: number of samples}

\item{lag}{numeric: fixed exposure period}
}
\value{
numeric list of posterior samples for infection and removal rates
}
\description{
This function extends \code{bayes_complete()} to the multi-type setting where there are different
infection and removal rates for different classes of individuals.
}
\details{
The infections vector is of the same length as the population size (number of individuals).
Note that this vector size differs from \code{bayes_complete()}, where the infections vector
is of the same length as the epidemic size (number of infected individuals).
The removals vector is of the same length as the epidemic size (number of infected individuals).
The classes are specified by \code{infection_classes} and \code{removal_classes}, which should be
integer vectors of the same length as \code{infections} and \code{removals}, respectively.
The rate priors for infection and removal rates are calculated based on initial estimates and shape priors.
The rate priors are divided by the population size for infection rates.
The priors and initial estimates should be provided as vectors
of the same length as the number of unique classes.
}
\keyword{internal}
