% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_complete_spatial.R
\name{bayes_complete_spatial}
\alias{bayes_complete_spatial}
\title{Posterior sampling for spatially-varying infection and removal rates given complete data}
\usage{
bayes_complete_spatial(
  removals,
  infections,
  population_size,
  kernel_spatial,
  matrix_distance,
  beta_init = 1,
  gamma_init = 1,
  beta_shape = 1,
  gamma_shape = 1,
  num_iter = 10000,
  num_renewals = 1,
  lag = 0
)
}
\arguments{
\item{removals}{numeric vector: removal times}

\item{infections}{numeric vector: infection times}

\item{population_size}{integer: population size}

\item{kernel_spatial}{function: symmetric function of distance}

\item{matrix_distance}{numeric: two-dimensional distance matrix}

\item{beta_init}{numeric: initial infection rate estimate}

\item{gamma_init}{numeric: initial removal rate estimate}

\item{beta_shape}{numeric: shape of gamma distribution prior}

\item{gamma_shape}{numeric: shape of gamma distribution prior}

\item{num_iter}{numeric: number of samples}

\item{lag}{numeric: fixed exposure period}
}
\value{
numeric list of posterior samples for infection and removal rates
}
\description{
This function extends \code{bayes_complete()} to the spatial setting where
infection rates vary by distance between individuals.
}
\details{
The spatial kernel function modulates the infection rate based on distance between individuals.
The infections vector is of the same length as the epidemic size (number of infected individuals).
The rate priors for infection and removal rates are calculated based on initial estimates and shape priors.
The rate priors are divided by the population size for infection rates.
}
\keyword{internal}
