% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomplete_sem.R
\name{decomplete_gsem}
\alias{decomplete_gsem}
\title{Missing epidemic data}
\usage{
decomplete_gsem(epi, p, q = 1)
}
\arguments{
\item{epi}{infection and removal times}

\item{p}{expected proportion of complete pairs observed}

\item{q}{probability infection time missing}
}
\value{
matrix: infection times, removal times
}
\description{
Impute NAs for infection and removal times
}
