% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomplete_sem.R
\name{decomplete_sem}
\alias{decomplete_sem}
\title{Missing epidemic data}
\usage{
decomplete_sem(epi, p, q = 1)
}
\arguments{
\item{epi}{infection and removal times}

\item{p}{expected proportion of complete pairs observed}

\item{q}{probability infection time missing}
}
\value{
matrix: infection times, removal times, (optional: infection claseses)
}
\description{
Impute NAs for infection and removal times
}
