% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sem.R
\name{filter_gsem}
\alias{filter_gsem}
\title{Filter epidemic data}
\usage{
filter_gsem(epi)
}
\arguments{
\item{epi}{matrix: infection times, removal times}
}
\value{
matrix: infection times, removal times
}
\description{
Keep cases only.
}
