% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_complete_data.R
\name{mle_complete_data}
\alias{mle_complete_data}
\title{MLE for complete stochastic epidemic model}
\usage{
mle_complete_data(r, i, N)
}
\arguments{
\item{r}{numeric vector: removal times}

\item{i}{numeric vector: infection times}

\item{N}{integer: population size}
}
\value{
MLEs for (infection.rate, removal.rate, R0)
}
\description{
Estimate infection and removal rates with complete epidemic observations.
}
