% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_removal_rate.R
\name{mle_removal_rate}
\alias{mle_removal_rate}
\title{MLE for Removal Rate}
\usage{
mle_removal_rate(r, i)
}
\arguments{
\item{r}{numeric vector: removal times}

\item{i}{numeric vector: infection times}
}
\value{
float: Removal rate estimate
}
\description{
Compute the maximum likelihood estimate for the (global) removal rate.
}
