% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peirr_imputed.R
\name{peirr_imputed}
\alias{peirr_imputed}
\title{Pair-based estimator of common infection and removal rates with imputed times}
\usage{
peirr_imputed(removals, infections, population_size, lag = 0)
}
\arguments{
\item{removals}{numeric vector: removal times}

\item{infections}{numeric vector: infection times}

\item{population_size}{integer: population size}

\item{lag}{numeric: fixed exposure period}
}
\value{
numeric list (infection.rate, removal.rate, R0)
}
\description{
Estimate infection and removal rates with tau-based expectation-maximization.
The output value \code{tau.sum} is useful for debugging.
}
