% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peirr_informed_pbla.R
\name{peirr_informed_pbla}
\alias{peirr_informed_pbla}
\title{PEIRR likelihood estimator of common infection rate and estimated removal rate}
\usage{
peirr_informed_pbla(r, i, N)
}
\arguments{
\item{r}{numeric vector: removal times}

\item{i}{numeric vector: infection times}

\item{N}{integer: population size}
}
\value{
numeric list (infection.rate, removal.rate, R0, tau.sum)
}
\description{
Estimate removal rate with duration date and infection rate with PBLA
}
