% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peirr_pbla_both_rates_multitype.R
\name{peirr_pbla_both_rates_multitype}
\alias{peirr_pbla_both_rates_multitype}
\title{PBLA estimators of infection rates and removal rates}
\usage{
peirr_pbla_both_rates_multitype(r, cr, ci, m = 1, A = 1, lag = 0)
}
\arguments{
\item{r}{numeric vector: removal times}

\item{cr}{numeric vector: removal time classes}

\item{ci}{numeric vector: infection time classes}

\item{m}{positive integer shape}

\item{A}{integer patient zeros}

\item{lag}{numeric fixed lag}

\item{N}{integer: population size}
}
\value{
numeric list (infection.rates, removal.rates)
}
\description{
Estimate multiple infection and removal rates with PBLA
}
