% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peirr_pbla_infection_rate.R
\name{peirr_pbla_infection_rate}
\alias{peirr_pbla_infection_rate}
\title{PEIRR likelihood estimator of infection rate with MLE removal rate}
\usage{
peirr_pbla_infection_rate(
  removals,
  infections,
  population_size,
  num_renewals = 1,
  num_patient_zeros = 1,
  lag = 0,
  known_gamma = NULL
)
}
\arguments{
\item{removals}{numeric vector: removal times}

\item{infections}{numeric vector: infection times}

\item{population_size}{integer: population size}

\item{num_renewals}{positive integer shape}

\item{num_patient_zeros}{integer patient zeros}

\item{lag}{numeric fixed lag}

\item{known_gamma}{numeric: removal rate}
}
\value{
numeric list (infection_rate, removal_rate, effective_number)
}
\description{
Estimate removal rate with duration date and infection rate with PBLA
}
