% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peirr_pbla_infection_rate_multitype.R
\name{peirr_pbla_infection_rate_multitype}
\alias{peirr_pbla_infection_rate_multitype}
\title{PEIRR likelihood estimator of infection rates and MLE removal rates}
\usage{
peirr_pbla_infection_rate_multitype(
  r,
  i,
  cr,
  ci,
  m = 1,
  A = 1,
  lag = 0,
  known.gammas = NULL,
  gamma.med = FALSE
)
}
\arguments{
\item{r}{numeric vector: removal times}

\item{i}{numeric vector: infection times}

\item{cr}{numeric vector: removal time classes}

\item{ci}{numeric vector: infection time classes}

\item{m}{positive integer shape}

\item{A}{integer patient zeros}

\item{lag}{numeric fixed lag}

\item{known.gammas}{numeric: removal rates}

\item{gamma.med}{bool: TRUE for median, and FALSE for mean in estimating the removal rate}

\item{N}{integer: population size}
}
\value{
numeric list (infection.rates, removal.rates)
}
\description{
Estimate multiple removal rates with duration date and multiple infection rates with PBLA
}
