% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peirr_tau.R
\name{peirr_tau}
\alias{peirr_tau}
\title{Pair-based tau estimator of common infection and removal rates}
\usage{
peirr_tau(r, i, N, lag = 0, tau.med = TRUE, gamma.med = FALSE)
}
\arguments{
\item{r}{numeric vector: removal times}

\item{i}{numeric vector: infection times}

\item{N}{integer: population size}

\item{lag}{numeric: fixed exposure period}

\item{tau.med}{bool: use median imputation for tau if TRUE}

\item{gamma.med}{bool: TRUE for median, and FALSE for mean in estimating the removal rate}
}
\value{
numeric list (infection.rate, removal.rate, R0)
}
\description{
Estimate infection and removal rates with tau-based expectation-maximization.
The output value \code{tau.sum} is useful for debugging.
}
