% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peirr_tau_seir.R
\name{peirr_tau_seir}
\alias{peirr_tau_seir}
\title{Pair-based tau estimator of common infection and removal rates with exposure delay}
\usage{
peirr_tau_seir(r, i, N, e = 0)
}
\arguments{
\item{r}{numeric vector: removal times}

\item{i}{numeric vector: infection times}

\item{N}{integer: population size}

\item{e}{numeric: fixed exposure period}
}
\value{
numeric list (infection.rate, removal.rate, R0, tau.sum)
}
\description{
Estimate infection and removal rates with tau-based expectation-maximization.
But, an individual has a period after exposure when they are not infectious.
The output value \code{tau.sum} is useful for debugging.
}
