% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peirr_tau_spatial.R
\name{peirr_tau_spatial}
\alias{peirr_tau_spatial}
\title{Pair-based tau estimator of common infection and removal rates with spatial effect}
\usage{
peirr_tau_spatial(
  removals,
  infections,
  population_size,
  kernel_spatial,
  matrix_distance,
  lag = 0,
  median_tau = FALSE,
  median_gamma = FALSE
)
}
\arguments{
\item{removals}{numeric: removal times}

\item{infections}{numeric: infection times}

\item{population_size}{integer: population size}

\item{kernel_spatial}{function: symmetric function of distance}

\item{matrix_distance}{numeric: two-dimensional distance matrix}

\item{lag}{numeric: fixed exposure period}

\item{median_tau}{bool: use median imputation for tau if TRUE}

\item{median_gamma}{bool: TRUE for median, and FALSE for mean in estimating the removal rate}
}
\value{
numeric list (infection.rate, removal.rate)
}
\description{
Estimate infection and removal rates with tau-based expectation-maximization.
The output value \code{tau.sum} is useful for debugging.
}
