% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_multitype_sem.R
\name{simulate_multitype_sem}
\alias{simulate_multitype_sem}
\title{Simulate general stochastic epidemic model with different infection rates}
\usage{
simulate_multitype_sem(betas, gamma, Ns, m = 1)
}
\arguments{
\item{betas}{numeric vector: infection rates}

\item{gamma}{numeric: removal rate}

\item{Ns}{integer: subpopulation sizes}

\item{m}{integer: positive shape}
}
\value{
matrix: infection times, removal times, infection classes
}
\description{
Draw infectious periods for general stochastic epidemic.
}
