% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_sem_multitype.R
\name{simulate_sem_multitype}
\alias{simulate_sem_multitype}
\title{Simulate general stochastic epidemic model with different infection rates}
\usage{
simulate_sem_multitype(betas, gamma, Ns, m = 1, e = 0)
}
\arguments{
\item{betas}{numeric vector: infection rates}

\item{gamma}{numeric: removal rate}

\item{Ns}{integer: subpopulation sizes}

\item{m}{integer: positive shape}

\item{e}{numeric: fixed exposure period}
}
\value{
numeric list: matrix of (infection times, removal times, classes), matrix of (St, It, Et, Rt, Time)
}
\description{
Draw infectious periods for general stochastic epidemic.
}
