% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_sem_spatial.R
\name{simulate_sem_spatial}
\alias{simulate_sem_spatial}
\title{Simulate stochastic epidemic model with spatial effect}
\usage{
simulate_sem_spatial(
  beta,
  gamma,
  population_size,
  kernel_spatial,
  matrix_distance,
  num_renewals = 1,
  lag = 0
)
}
\arguments{
\item{beta}{numeric: infection rate}

\item{gamma}{numeric: removal rate}

\item{population_size}{integer: population size}

\item{kernel_spatial}{function: symmetric function of distance}

\item{matrix_distance}{numeric: two-dimensional distance matrix}

\item{num_renewals}{integer: positive shape}

\item{lag}{numeric: fixed exposure period}
}
\value{
numeric list: matrix of (infection times, removal times, spatial coordinates), matrix of (St, It, Et, Rt, Time), matrix of N by N distances
}
\description{
Draw infectious periods for stochastic epidemic with spatial effect.
}
