% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_sem.R
\name{sort_sem}
\alias{sort_sem}
\title{Sort epidemic data}
\usage{
sort_sem(epi)
}
\arguments{
\item{epi}{matrix: infection times, removal times}
}
\value{
matrix: infection times, removal times
}
\description{
Sort matrix by increasing removal times.
}
