% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_complete.R
\name{bayes_complete}
\alias{bayes_complete}
\title{Posterior sampling for infection and removal rates given complete data}
\usage{
bayes_complete(
  removals,
  infections,
  population_size,
  beta_init = 1,
  gamma_init = 1,
  beta_shape = 1,
  gamma_shape = 1,
  num_iter = 10000,
  num_renewals = 1,
  lag = 0
)
}
\arguments{
\item{removals}{numeric vector: removal times}

\item{infections}{numeric vector: infection times}

\item{population_size}{integer: population size}

\item{beta_init}{numeric: initial infection rate estimate}

\item{gamma_init}{numeric: initial removal rate estimate}

\item{beta_shape}{numeric: shape of gamma distribution prior}

\item{gamma_shape}{numeric: shape of gamma distribution prior}

\item{num_iter}{numeric: number of samples}

\item{lag}{numeric: fixed exposure period}
}
\value{
numeric list of posterior samples for infection and removal rates
}
\description{
This is the base function for Gibbs sampling of infection and removal rates
when infection times and removal times are fully observed.
}
\details{
This function assumes that the infections and removals vectors contain complete data
and are of the same length as the epidemic size (number of infected individuals).
The rate priors for infection and removal rates are calculated based on initial estimates and shape priors.
The rate prior is divided by the population size for infection rate.
}
\keyword{internal}
