% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_complete_data.R
\name{bayes_complete_data}
\alias{bayes_complete_data}
\title{Posterior parameters for infection and removal rates given complete data}
\usage{
bayes_complete_data(
  r,
  i,
  N,
  infection.rate.rate.prior = 1e-05,
  infection.rate.shape.prior = 0.001,
  removal.rate.rate.prior = 0.001,
  removal.rate.shape.prior = 0.001,
  num.posterior.samples = 10000
)
}
\arguments{
\item{r}{numeric vector: removal times}

\item{i}{numeric vector: infection times}

\item{N}{integer: population size}

\item{infection.rate.rate.prior}{numeric}

\item{infection.rate.shape.prior}{numeric}

\item{removal.rate.rate.prior}{numeric}

\item{removal.rate.shape.prior}{numeric}

\item{num.posterior.samples}{numeric}
}
\value{
numeric list of exact posterior and prior parameters
}
\description{
Parameters for independent Gibbs sampling of the infection and removal rates.
You can sample from \code{rgamma()} with the posterior parameters to get the posterior distribution.
For the infection rate, you should scale the result of \code{rgamma()} by the population size.
\code{infection.rate.samples} and \code{removal.rate.samples} contain the posterior samples.
}
