% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_complete_spatial.R
\name{bayes_complete_spatial}
\alias{bayes_complete_spatial}
\title{Posterior parameters for spatially-varying infection and removal rates given complete data}
\usage{
bayes_complete_spatial(
  removals,
  infections,
  population_size,
  kernel_spatial,
  matrix_distance,
  beta_init = 1,
  gamma_init = 1,
  beta_shape = 1,
  gamma_shape = 1,
  num_iter = 10000,
  num_renewals = 1,
  lag = 0
)
}
\arguments{
\item{removals}{numeric vector: removal times}

\item{infections}{numeric vector: infection times}

\item{population_size}{integer: population size}

\item{kernel_spatial}{function: symmetric function of distance}

\item{matrix_distance}{numeric: two-dimensional distance matrix}

\item{beta_init}{numeric}

\item{gamma_init}{numeric}

\item{beta_shape}{numeric}

\item{gamma_shape}{numeric}

\item{num_iter}{numeric}

\item{lag}{numeric: fixed exposure period}
}
\value{
numeric list of exact posterior and prior parameters
}
\description{
Parameters for independent Gibbs sampling of the infection and removal rates.
You can sample from \code{rgamma()} with the posterior parameters to get the posterior distribution.
For the infection rate, you should scale the result of \code{rgamma()} by the population size.
\code{infection.rate.samples} and \code{removal.rate.samples} contain the posterior samples.
}
\keyword{internal}
