% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomplete_sem.R
\name{decomplete_sem}
\alias{decomplete_sem}
\title{Missing epidemic data}
\usage{
decomplete_sem(epidemic_time, prop_complete = 0, prop_infection_missing = 1)
}
\arguments{
\item{epidemic_time}{matrix: infection and removal times}

\item{prop_complete}{numeric: expected proportion of complete pairs observed}

\item{prop_infection_missing}{numeric: probability infection time missing}
}
\value{
matrix: infection times, removal times, (optional: infection classes)
}
\description{
Impute NAs for infection and removal times
}
