% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sem.R
\name{filter_sem}
\alias{filter_sem}
\title{Filter epidemic data}
\usage{
filter_sem(epi)
}
\arguments{
\item{epi}{matrix: infection times, removal times, (optional: infection classes)}
}
\value{
matrix: infection times, removal times, (optional: infection classes)
}
\description{
Keep cases only.
}
