% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peirr_bayes.R
\name{peirr_bayes}
\alias{peirr_bayes}
\title{Posterior samples for infection and removal rates given incomplete data}
\usage{
peirr_bayes(r, i, N)
}
\arguments{
\item{r}{numeric vector: removal times}

\item{i}{numeric vector: infection times}

\item{N}{integer: population size}
}
\value{
numeric list of posterior (infection.rate, removal.rate)
}
\description{
Gibbs sample the infection and removal rates
}
