% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peirr_pbla_both_rates.R
\name{peirr_pbla_both_rates}
\alias{peirr_pbla_both_rates}
\title{PBLA estimator of common infection rate and estimated removal rate}
\usage{
peirr_pbla_both_rates(r, N)
}
\arguments{
\item{r}{numeric vector: removal times}

\item{N}{integer: population size}
}
\value{
numeric list (infection.rate, removal.rate, R0, tau.sum)
}
\description{
Estimate infection and removal rates with PBLA
}
