% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peirr_pbla_both_rates.R
\name{peirr_pbla_both_rates}
\alias{peirr_pbla_both_rates}
\title{PBLA estimator of infection rate and removal rate}
\usage{
peirr_pbla_both_rates(
  removals,
  population_size,
  num_renewals = 1,
  num_patient_zeros = 1,
  lag = 0,
  infections = NA
)
}
\arguments{
\item{removals}{numeric: removal times}

\item{population_size}{integer: population size}

\item{num_renewals}{integer: erlang shape}

\item{num_patient_zeros}{integer: patient zeros}

\item{lag}{numeric: fixed lag}
}
\value{
numeric list (infection_rate, removal_rate, effective_number)
}
\description{
Estimate infection and removal rates with PBLA
}
