% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peirr_pbla_both_rates_multitype.R
\name{peirr_pbla_both_rates_multitype}
\alias{peirr_pbla_both_rates_multitype}
\title{PBLA joint estimator of infection rates and removal rates}
\usage{
peirr_pbla_both_rates_multitype(
  removals,
  removal_classes,
  infection_classes,
  num_renewals = 1,
  num_patient_zeros = 1,
  lag = 0
)
}
\arguments{
\item{removals}{numeric: removal times (as large a vector as population size, with NAs for uninfected)}

\item{removal_classes}{numeric: removal time classes (as large a vector as population size, with NAs for uninfected)}

\item{infection_classes}{numeric: infection time classes (as large a vector as population size, with NAs for uninfected)}

\item{num_renewals}{integer: erlang shape}

\item{num_patient_zeros}{integer: patient zeros}

\item{lag}{numeric: fixed lag}
}
\value{
numeric list (infection_rate, removal_rate)
}
\description{
Joint estimate multiple infection and removal rates with PBLA
}
