% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peirr_pbla_infection_rate.R
\name{peirr_pbla_infection_rate}
\alias{peirr_pbla_infection_rate}
\title{PEIRR likelihood estimator of common infection rate and estimated removal rate}
\usage{
peirr_pbla_infection_rate(r, i, N, m = 1, A = 1, lag = 0, known.gamma = NULL)
}
\arguments{
\item{r}{numeric vector: removal times}

\item{i}{numeric vector: infection times}

\item{N}{integer: population size}

\item{m}{positive integer shape}

\item{A}{integer patient zeros}

\item{lag}{numeric fixed lag}

\item{known.gamma}{numeric: removal rate}
}
\value{
numeric list (infection.rate, removal.rate, R0, tau.sum)
}
\description{
Estimate removal rate with duration date and infection rate with PBLA
}
