% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peirr_pbla_infection_rate_multitype.R
\name{peirr_pbla_infection_rate_multitype}
\alias{peirr_pbla_infection_rate_multitype}
\title{PBLA estimator of infection rates and MLE removal rates}
\usage{
peirr_pbla_infection_rate_multitype(
  removals,
  infections,
  removal_classes,
  infection_classes,
  num_renewals = 1,
  num_patient_zeros = 1,
  lag = 0,
  known_gamma = NULL
)
}
\arguments{
\item{removals}{numeric: removal times (as large a vector as population size, with NAs for uninfected)}

\item{infections}{numeric: infection times (as large a vector as population size, with NAs for uninfected)}

\item{removal_classes}{numeric: removal time classes (as large a vector as population size, with NAs for uninfected)}

\item{infection_classes}{numeric: infection time classes (as large a vector as population size, with NAs for uninfected)}

\item{num_renewals}{integer: erlang shape}

\item{num_patient_zeros}{integer: patient zeros}

\item{lag}{numeric: fixed lag}

\item{known_gamma}{numeric: removal rates}
}
\value{
numeric list (infection_rate, removal_rate)
}
\description{
Estimate multiple removal rates with duration date and multiple infection rates with PBLA
}
