% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peirr_removal_rate.R
\name{peirr_removal_rate}
\alias{peirr_removal_rate}
\title{Removal rate estimator}
\usage{
peirr_removal_rate(r, i, med = TRUE)
}
\arguments{
\item{r}{numeric vector: removal times}

\item{i}{numeric vector: infection times}

\item{med}{bool: TRUE for median, FALSE for mean}
}
\value{
float: Removal rate estimate
}
\description{
Estimate the removal rate as the inverse of the mean or median infection period.
}
