% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peirr_removal_rate.R
\name{peirr_removal_rate}
\alias{peirr_removal_rate}
\title{Removal rate estimator}
\usage{
peirr_removal_rate(removals, infections)
}
\arguments{
\item{removals}{numeric vector: removal times}

\item{infections}{numeric vector: infection times}
}
\value{
float: Removal rate estimate
}
\description{
Estimate the removal rate as the inverse of the mean infectious period length.
}
