% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peirr_tau.R
\name{peirr_tau}
\alias{peirr_tau}
\title{Pair-based tau estimator of common infection and removal rates}
\usage{
peirr_tau(
  removals,
  infections,
  population_size,
  lag = 0,
  median_tau = FALSE,
  median_gamma = FALSE
)
}
\arguments{
\item{removals}{numeric vector: removal times}

\item{infections}{numeric vector: infection times}

\item{population_size}{integer: population size}

\item{lag}{numeric: fixed exposure period}

\item{median_tau}{bool: use median imputation for tau if TRUE}

\item{median_gamma}{bool: TRUE for median, and FALSE for mean in estimating the removal rate}
}
\value{
numeric list (infection.rate, removal.rate, R0)
}
\description{
Estimate infection and removal rates with tau-based expectation-maximization.
The output value \code{tau.sum} is useful for debugging.
}
