% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peirr_tau.R
\name{peirr_tau}
\alias{peirr_tau}
\title{Pair-based tau estimator of common infection and removal rates}
\usage{
peirr_tau(r, i, N)
}
\arguments{
\item{r}{numeric vector: removal times}

\item{i}{numeric vector: infection times}

\item{N}{integer: population size}
}
\value{
numeric list (infection.rate, removal.rate, R0, tau.sum)
}
\description{
Estimate infection and removal rates with tau-based expectation-maximization.
The output value \code{tau.sum} is useful for debugging.
}
