% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peirr_tau_multitype.R
\name{peirr_tau_multitype}
\alias{peirr_tau_multitype}
\title{Pair-based tau estimator of multiple infection and removal rates}
\usage{
peirr_tau_multitype(
  r,
  i,
  cr,
  ci,
  Ns,
  lag = 0,
  tau.med = TRUE,
  gamma.med = FALSE
)
}
\arguments{
\item{r}{numeric vector: removal times}

\item{i}{numeric vector: infection times}

\item{cr}{numeric vector: removal time classes}

\item{ci}{numeric vector: infection time classes}

\item{Ns}{integer vector: population sizes for infection classes (sorted)}

\item{lag}{numeric: fixed exposure period}

\item{tau.med}{bool: use median imputation for tau if TRUE}

\item{gamma.med}{bool: TRUE for median, and FALSE for mean in estimating the removal rate}
}
\value{
numeric list (infection.rates, removal.rates, removal.full.sizes)
}
\description{
Estimate infection and removal rates with tau-based expectation-maximization.
The output value \code{tau.sum} is useful for debugging.
}
