% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peirr_tau_multitype.R
\name{peirr_tau_multitype}
\alias{peirr_tau_multitype}
\title{Pair-based tau estimator of multiple infection and removal rates}
\usage{
peirr_tau_multitype(
  removals,
  infections,
  removal_classes,
  infection_classes,
  infection_class_sizes,
  lag = 0
)
}
\arguments{
\item{removals}{numeric: removal times}

\item{infections}{numeric: infection times}

\item{removal_classes}{numeric: removal time classes}

\item{infection_classes}{numeric: infection time classes}

\item{infection_class_sizes}{integer: population sizes for infection classes (sorted)}

\item{lag}{numeric: fixed exposure period}
}
\value{
numeric list (infection_rates, removal_rates, etc)
}
\description{
Estimate infection and removal rates with tau-based expectation-maximization.
The output value \code{tau.sum} is useful for debugging.
}
