% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peirr_tau_spatial.R
\name{peirr_tau_spatial}
\alias{peirr_tau_spatial}
\title{Pair-based tau estimator of common infection and removal rates with spatial effect}
\usage{
peirr_tau_spatial(r, i, N, h, D, lag = 0, tau.med = FALSE, gamma.med = FALSE)
}
\arguments{
\item{r}{numeric vector: removal times}

\item{i}{numeric vector: infection times}

\item{N}{integer: population size}

\item{h}{function: symmetric function of distance}

\item{D}{numeric: two-dimensional distance matrix}

\item{lag}{numeric: fixed exposure period}

\item{tau.med}{bool: use median imputation for tau if TRUE}

\item{gamma.med}{bool: TRUE for median, and FALSE for mean in estimating the removal rate}
}
\value{
numeric list (infection.rate, removal.rate)
}
\description{
Estimate infection and removal rates with tau-based expectation-maximization.
The output value \code{tau.sum} is useful for debugging.
}
