% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_sem.R
\name{simulate_sem}
\alias{simulate_sem}
\title{Simulate general stochastic epidemic model}
\usage{
simulate_sem(beta, gamma, population_size, num_renewals = 1, lag = 0)
}
\arguments{
\item{beta}{numeric: infection rate}

\item{gamma}{numeric: removal rate}

\item{population_size}{integer: population size}

\item{num_renewals}{integer: positive shape}

\item{lag}{numeric: fixed exposure period}
}
\value{
numeric list: matrix of (infection times, removal times), matrix of (St, It, Et, Rt, Time)
}
\description{
Draw infectious periods for general stochastic epidemic.
}
\keyword{internal}
