% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_sem.R
\name{simulate_sem}
\alias{simulate_sem}
\title{Simulate general stochastic epidemic model}
\usage{
simulate_sem(beta, gamma, N, m = 1, e = 0)
}
\arguments{
\item{beta}{numeric: infection rate}

\item{gamma}{numeric: removal rate}

\item{N}{integer: population size}

\item{m}{integer: positive shape}

\item{e}{numeric: fixed exposure period}
}
\value{
numeric list: matrix of (infection times, removal times), matrix of (St, It, Et, Rt, Time)
}
\description{
Draw infectious periods for general stochastic epidemic.
}
