% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_sem_multitype.R
\name{simulate_sem_multitype}
\alias{simulate_sem_multitype}
\title{Simulate general stochastic epidemic model with different infection rates}
\usage{
simulate_sem_multitype(
  beta,
  gamma,
  infection_class_sizes,
  removal_class_sizes,
  num_renewals = 1,
  lag = 0
)
}
\arguments{
\item{beta}{numeric vector: infection rates}

\item{gamma}{numeric: removal rates}

\item{infection_class_sizes}{integer: subpopulation sizes for infection rates}

\item{removal_class_sizes}{integer: subpopulation sizes for removal rates}

\item{num_renewals}{integer: positive erlang shape}

\item{lag}{numeric: fixed exposure period}
}
\value{
numeric list: matrix of (infection times, removal times, classes and rates), matrix of (St, It, Et, Rt, Time)
}
\description{
Draw infectious periods for general stochastic epidemic.
}
