% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_sem_multitype.R
\name{simulate_sem_multitype}
\alias{simulate_sem_multitype}
\title{Simulate general stochastic epidemic model with different infection rates}
\usage{
simulate_sem_multitype(betas, gammas, beta.sizes, gamma.sizes, m = 1, lag = 0)
}
\arguments{
\item{betas}{numeric vector: infection rates}

\item{gammas}{numeric: removal rates}

\item{beta.sizes}{integers: subpopulation sizes for infection rates}

\item{gamma.sizes}{integers: subpopulation sizes for removal rates}

\item{m}{integer: positive shape}

\item{lag}{numeric: fixed exposure period}
}
\value{
numeric list: matrix of (infection times, removal times, classes and rates), matrix of (St, It, Et, Rt, Time)
}
\description{
Draw infectious periods for general stochastic epidemic.
}
