% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_sem_spatial.R
\name{simulate_sem_spatial}
\alias{simulate_sem_spatial}
\title{Simulate stochastic epidemic model with spatial effect}
\usage{
simulate_sem_spatial(beta, gamma, N, h, D, m = 1, lag = 0)
}
\arguments{
\item{beta}{numeric: infection rate}

\item{gamma}{numeric: removal rate}

\item{N}{integer: population size}

\item{h}{function: symmetric function of distance}

\item{D}{numeric: two-dimensional distance matrix}

\item{m}{integer: positive shape}

\item{lag}{numeric: fixed exposure period}
}
\value{
numeric list: matrix of (infection times, removal times, spatial coordinates), matrix of (St, It, Et, Rt, Time), matrix of N by N distances
}
\description{
Draw infectious periods for stochastic epidemic with spatial effect.
}
