% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulator}
\alias{simulator}
\title{Simulate general stochastic epidemic model with formatting}
\usage{
simulator(beta, gamma, N, m = 1, e = 0, p = 0, q = 1)
}
\arguments{
\item{beta}{numeric: infection rate}

\item{gamma}{numeric: removal rate}

\item{N}{integer: population size}

\item{m}{integer: positive shape}

\item{e}{numeric: fixed exposure period}

\item{p}{expected proportion of complete pairs observed}

\item{q}{probability infection time missing}
}
\value{
numeric list: matrix of (infection times, removal times), matrix of (St, It, Et, Rt, Time)
}
\description{
Draw infectious periods for general stochastic epidemic.
}
