% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{simulator}
\alias{simulator}
\title{Simulate general stochastic epidemic model with formatting}
\usage{
simulator(
  beta,
  gamma,
  N,
  m = 1,
  lag = 0,
  p = 0,
  q = 1,
  min.sample.size = 10,
  max.sample.size = Inf
)
}
\arguments{
\item{beta}{numeric: infection rate}

\item{gamma}{numeric: removal rate}

\item{N}{integer: population size}

\item{m}{integer: positive shape}

\item{lag}{numeric: fixed exposure period}

\item{p}{expected proportion of complete pairs observed}

\item{q}{probability infection time missing}

\item{min.sample.size}{integer}
}
\value{
numeric list: matrix of (infection times, removal times), matrix of (St, It, Et, Rt, Time)
}
\description{
Draw infectious periods for general stochastic epidemic.
}
