% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator_multitype.R
\name{simulator_multitype}
\alias{simulator_multitype}
\title{Simulate multi-type stochastic epidemic model with formatting}
\usage{
simulator_multitype(
  betas,
  gammas,
  beta.sizes,
  gamma.sizes,
  m = 1,
  lag = 0,
  p = 0.5,
  q = 1,
  min.sample.size = 10,
  max.sample.size = Inf
)
}
\arguments{
\item{betas}{numeric vector: infection rates}

\item{beta.sizes}{integers: subpopulation sizes for infection rates}

\item{gamma.sizes}{integers: subpopulation sizes for removal rates}

\item{m}{integer: positive shape}

\item{lag}{numeric: fixed exposure period}

\item{p}{expected proportion of complete pairs observed}

\item{q}{probability infection time missing}

\item{min.sample.size}{integer}

\item{gamma}{numeric: removal rates}
}
\value{
numeric list: matrix of (infection times, removal times), matrix of (St, It, Et, Rt, Time)
}
\description{
Draw infectious periods for stochastic epidemic with different classes.
}
