% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator_multitype.R
\name{simulator_multitype}
\alias{simulator_multitype}
\title{Simulate multi-type stochastic epidemic model with formatting}
\usage{
simulator_multitype(
  betas,
  gammas,
  infection_class_sizes,
  removal_class_sizes,
  num_renewals = 1,
  lag = 0,
  prop_complete = 0.5,
  prop_infection_missing = 1,
  min_epidemic_size = 10,
  max_epidemic_size = Inf
)
}
\arguments{
\item{betas}{numeric vector: infection rates}

\item{gammas}{numeric vector: removal rates}

\item{infection_class_sizes}{integers: subpopulation sizes for infection rates}

\item{removal_class_sizes}{integers: subpopulation sizes for removal rates}

\item{num_renewals}{integer: positive shape}

\item{lag}{numeric: fixed exposure period}

\item{prop_complete}{expected proportion of complete pairs observed}

\item{prop_infection_missing}{probability infection time missing}

\item{min_epidemic_size}{integer}

\item{max_epidemic_size}{integer}
}
\value{
numeric list: matrix of (infection times, removal times), matrix of (St, It, Et, Rt, Time)
}
\description{
Draw infectious periods for stochastic epidemic with different classes.
}
