% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator_spatial.R
\name{simulator_spatial}
\alias{simulator_spatial}
\title{Simulate spatial stochastic epidemic model with formatting}
\usage{
simulator_spatial(
  beta,
  gamma,
  population_size,
  kernel,
  matrix_distance,
  num_renewals = 1,
  lag = 0,
  prop_complete = 0.5,
  prop_infection_missing = 1,
  min_epidemic_size = 10,
  max_epidemic_size = Inf
)
}
\arguments{
\item{beta}{numeric: infection rate}

\item{gamma}{numeric: removal rate}

\item{population_size}{integer: population size}

\item{kernel}{function: symmetric function of distance}

\item{matrix_distance}{numeric: two-dimensional distance matrix}

\item{num_renewals}{integer: positive shape}

\item{lag}{numeric: fixed exposure period}

\item{prop_complete}{expected proportion of complete pairs observed}

\item{prop_infection_missing}{probability infection time missing}

\item{min_epidemic_size}{integer}

\item{max_epidemic_size}{integer}
}
\value{
numeric list: matrix of (infection times, removal times), matrix of (St, It, Et, Rt, Time), matrix of n by N distances
}
\description{
Draw infectious periods for spatial stochastic epidemic.
}
