% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator_spatial.R
\name{simulator_spatial}
\alias{simulator_spatial}
\title{Simulate spatial stochastic epidemic model with formatting}
\usage{
simulator_spatial(
  beta,
  gamma,
  N,
  h,
  D,
  m = 1,
  lag = 0,
  p = 0.5,
  q = 1,
  min.sample.size = 10,
  max.sample.size = Inf
)
}
\arguments{
\item{beta}{numeric: infection rate}

\item{gamma}{numeric: removal rate}

\item{N}{integer: population size}

\item{h}{function: symmetric function of distance}

\item{D}{numeric: two-dimensional distance matrix}

\item{m}{integer: positive shape}

\item{lag}{numeric: fixed exposure period}

\item{p}{expected proportion of complete pairs observed}

\item{q}{probability infection time missing}

\item{min.sample.size}{integer}
}
\value{
numeric list: matrix of (infection times, removal times), matrix of (St, It, Et, Rt, Time), matrix of n by N distances
}
\description{
Draw infectious periods for spatial stochastic epidemic.
}
