% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_sem.R
\name{sort_sem}
\alias{sort_sem}
\title{Sort epidemic data}
\usage{
sort_sem(matrix_time)
}
\arguments{
\item{matrix_time}{matrix: infection times, removal times, (optional: infection classes)}
}
\value{
matrix: infection times, removal times, (optional: infection classes)
}
\description{
Sort matrix by increasing removal times.
}
\keyword{internal}
