% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_moment.R
\name{tau_moment}
\alias{tau_moment}
\title{Expected value for tau with different missingness}
\usage{
tau_moment(rk, rj, ik, ij, lambdak, lambdaj, med = T)
}
\arguments{
\item{rk}{numeric: removal time rk}

\item{rj}{numeric: removal time rj}

\item{ik}{numeric: infection time ik}

\item{ij}{numeric: infection time ij}

\item{lambdak}{numeric: removal rate of k}

\item{lambdaj}{numeric: removal rate of j}

\item{med}{bool: use median imputation if true}
}
\value{
numeric: expectation of some pair-wise tau
}
\description{
Compute expectation of tau with (in)complete observations.
}
